deploy

+++ safeAI-plugin/src/test/java/com/safeai/neo4jplugin/GovernanceModuleTest.java
@@ -1,8 +1,9 @@
 package com.safeai.neo4jplugin;
 
-import com.safeai.neo4jplugin.governance.GovernanceProcedures;
-import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.jupiter.api.Test;
+
+import com.safeai.neo4jplugin.governance.GovernanceProcedures;
 
 /**
  * Unit tests for GovernanceProcedures.
@@ -12,8 +13,8 @@ public class GovernanceModuleTest {
     public void testGovernanceVoting() {
         String proposalId = "proposal123";
         GovernanceProcedures governance = new GovernanceProcedures();
-        int initialVotes = governance.initiateVote(proposalId);
-        int updatedVotes = governance.recordVote(proposalId, 1);
+        long initialVotes = governance.initiateVote(proposalId);
+        long updatedVotes = governance.recordVote(proposalId, 1);
         assertEquals(initialVotes + 1, updatedVotes);
     }
 }
